トランプ機能の実装
５２枚の中からランダムでカードを取得するトランプを実装しました

+++ Utility.h
@@ -4,4 +4,4 @@
 #define TRUE 0
 #define FALSE -1
 
-typedef short int BOOL;
+typedef int BOOL;
